Move write_to_file to a tool file (#2093)
